Revert "feat: AdminLessonInputs support challenges"
This reverts commit 58fa9b2e49ce37d47277579ce8fda4b000837e06.
