--- components/admin/lessons/AdminLessonInputs/index.tsx
@@ -1,1 +1,0 @@
-export { default } from './AdminLessonInputs'
